AnyScript - at least build works
