+++ index.html
@@ -9,6 +9,6 @@
   <body>
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
-    <script type="module" src="/src/main.js"></script>
+    <script type="module" src="/src/tmws.ts"></script>
   </body>
 </html>
